2019/01/28
튜토리얼에서 조작화면이 나오도록 변경

+++ Task_Tutorial.h
@@ -57,7 +57,7 @@ namespace Tutorial
 		int countDown;
 		bool countDown_Flag;
 
-		int timeCnt;
+		float timeCnt;
 		bool page_Change_Flag;
 		int page_Change_Speed;
